Matching Game
Part 1. Most fundamental layout
html; css
Part 2. Generating the left side images
 Generate the faces on the left side
Bind all the properties and functions with a global object in the <Script>
Part 3. Extend the JavaScript code developed in Part 2 to handle the right side
 Copy faces from left side to the right side
added copyFacesToRight() method
Part 4. Finishing the game
 Add the click event handler
added event handler for the lastChild for the left side and the whole body